+++ src/player/rusty_backend/sink.rs
@@ -271,9 +271,10 @@ impl Sink {
         if let Some(sleep_until_end) = self.sleep_until_end.lock().unwrap().take() {
             std::thread::spawn(move || {
                 let _ = sleep_until_end.recv();
-                if let Err(e) = tx1.send(PlayerMsg::Eos) {
-                    eprintln!("Error is: {}", e);
-                }
+                tx1.send(PlayerMsg::Eos).ok();
+                // if let Err(e) = tx1.send(PlayerMsg::Eos) {
+                //     eprintln!("Error is: {}", e);
+                // }
             });
         }
     }